requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.14.0
python-dotenv==1.0.0